🐞 fix: avoid processing all targets in code fence
🐞 fix: avoid processing target in code fence

+++ src/options/addImgCaptionText.ts
@@ -1,10 +1,20 @@
 import nlToolsKit from "main";
 import { deleteArrayElement, fileContentsProcess, processFunc } from "../utils";
+import {
+	endMardownCodefencesymbol,
+	startMardownCodefencesymbol,
+} from "src/config/regex";
 export const deleteImageCaptionText = new fileContentsProcess((lines) => {
+	let isIncode = false;
 	for (let i = 0; i < lines.length; i++) {
+		if (lines[i].match(startMardownCodefencesymbol)) {
+			isIncode = !isIncode;
+		} else if (lines[i].match(endMardownCodefencesymbol)) {
+			isIncode = !isIncode;
+		}
 		const targetDeletedImgCaptionStrRegex =
-			/<center>(图 |Figure )\d-\d-.*?<\/center>/;
-		if (lines[i].match(targetDeletedImgCaptionStrRegex)) {
+			/<center>(图 |Figure )\d+-\d+-.*?<\/center>/;
+		if (lines[i].match(targetDeletedImgCaptionStrRegex) && !isIncode) {
 			lines = deleteArrayElement(lines, i);
 		}
 	}
@@ -12,9 +22,16 @@ export const deleteImageCaptionText = new fileContentsProcess((lines) => {
 });
 export const addImgCaptionText = new fileContentsProcess((lines, plugin) => {
 	let imgCaptionNum = [0, 0];
+	let isIncode = false;
 	for (let i = 0; i < lines.length; i++) {
+		// Skip the situation that target within code fencce
+		if (lines[i].match(startMardownCodefencesymbol)) {
+			isIncode = !isIncode;
+		} else if (lines[i].match(endMardownCodefencesymbol)) {
+			isIncode = !isIncode;
+		}
 		const header2Regex = /^##( +).*\n?/m;
-		if (lines[i].match(header2Regex)) {
+		if (lines[i].match(header2Regex) && !isIncode) {
 			imgCaptionNum[0]++;
 			imgCaptionNum[1] = 0;
 		}
@@ -25,8 +42,9 @@ export const addImgCaptionText = new fileContentsProcess((lines, plugin) => {
 		// If the current line is not the line including image reference link, skip it
 
 		if (
-			lines[i].match(imgCaptionMdRegex) ||
-			lines[i].match(imgCaptionWikiRegex)
+			!isIncode &&
+			(lines[i].match(imgCaptionMdRegex) ||
+				lines[i].match(imgCaptionWikiRegex))
 		) {
 			lines[i] =
 				imgCaptionProcess(
